+++ tilequeue/format/mvt.py
@@ -3,6 +3,10 @@ from mapbox_vector_tile import encode as mvt_encode
 
 
 def encode(fp, feature_layers, coord, bounds_merc):
-    tile = mvt_encode(feature_layers, quantize_bounds=bounds_merc,
-                      on_invalid_geometry=on_invalid_geometry_make_valid)
+    tile = mvt_encode(
+        feature_layers,
+        quantize_bounds=bounds_merc,
+        on_invalid_geometry=on_invalid_geometry_make_valid,
+        round_fn=round,
+    )
     fp.write(tile)